feat: support CPU Profile
