+++ pkg/cli/runner.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"io"
 	"os"
-	"runtime/trace"
 	"time"
 
 	"github.com/aquaproj/aqua/pkg/config"
@@ -83,6 +82,10 @@ func (runner *Runner) Run(ctx context.Context, args ...string) error {
 				Name:  "trace",
 				Usage: "trace output file path",
 			},
+			&cli.StringFlag{
+				Name:  "cpu-profile",
+				Usage: "cpu profile output file path",
+			},
 		},
 		Commands: []*cli.Command{
 			runner.newInstallCommand(),
@@ -103,32 +106,4 @@ func exitErrHandlerFunc(c *cli.Context, err error) {
 		cli.HandleExitCoder(err)
 		return
 	}
-}
-
-type Tracer struct {
-	f io.Closer
-}
-
-func (t *Tracer) Stop() {
-	if t == nil {
-		return
-	}
-	trace.Stop()
-}
-
-func startTrace(p string) (*Tracer, error) {
-	if p == "" {
-		return nil, nil //nolint:nilnil
-	}
-	f, err := os.Create(p)
-	if err != nil {
-		return nil, fmt.Errorf("create a trace output file: %w", err)
-	}
-	if err := trace.Start(f); err != nil {
-		f.Close()
-		return nil, fmt.Errorf("start a trace: %w", err)
-	}
-	return &Tracer{
-		f: f,
-	}, nil
 }
