+++ pkg/cli/which.go
@@ -40,6 +40,13 @@ func (runner *Runner) whichAction(c *cli.Context) error {
 		return err
 	}
 	defer tracer.Stop()
+
+	cpuProfiler, err := startCPUProfile(c.String("cpu-profile"))
+	if err != nil {
+		return err
+	}
+	defer cpuProfiler.Stop()
+
 	param := &config.Param{}
 	if err := runner.setParam(c, param); err != nil {
 		return fmt.Errorf("parse the command line arguments: %w", err)
